+++ src/browser/settings.jsx
@@ -1,6 +1,6 @@
 'use strict';
 
-const remote = require('electron').remote;
+const {remote, ipcRenderer} = require('electron');
 const settings = require('../common/settings');
 
 const React = require('react');
@@ -34,9 +34,7 @@ var SettingsPage = React.createClass({
       config = settings.loadDefault();
     }
 
-    this.setState({
-      showAddTeamForm: false
-    });
+    config.showAddTeamForm = false;
 
     return config;
   },
@@ -54,10 +52,8 @@ var SettingsPage = React.createClass({
     this.setState({
       teams: teams
     });
-
-    this.handleSave(false);
   },
-  handleSave: function(toIndex) {
+  handleSave: function() {
     var config = {
       teams: this.state.teams,
       hideMenuBar: this.state.hideMenuBar,
@@ -85,9 +81,10 @@ var SettingsPage = React.createClass({
       });
     }
 
-    if (typeof toIndex == 'undefined' || toIndex) {
-      backToIndex();
-    }
+    ipcRenderer.send('update-menu', config);
+    ipcRenderer.send('update-config');
+
+    backToIndex();
   },
   handleCancel: function() {
     backToIndex();
@@ -117,16 +114,10 @@ var SettingsPage = React.createClass({
       autostart: this.refs.autostart.getChecked()
     });
   },
-  handleShowTeamForm: function() {
-    if (!this.state.showAddTeamForm) {
-      this.setState({
-        showAddTeamForm: true
-      });
-    } else {
-      this.setState({
-        showAddTeamForm: false
-      });
-    }
+  toggleShowTeamForm: function() {
+    this.setState({
+      showAddTeamForm: !this.state.showAddTeamForm
+    });
   },
   handleFlashWindowSetting: function(item) {
     this.setState({
@@ -136,11 +127,6 @@ var SettingsPage = React.createClass({
     });
   },
   render: function() {
-
-    var buttonStyle = {
-      marginTop: 20
-    };
-
     var teams_row = (
     <Row>
       <Col md={ 12 }>
@@ -166,7 +152,7 @@ var SettingsPage = React.createClass({
     }
     options.push(<Input key="inputDisableWebSecurity" ref="disablewebsecurity" type="checkbox" label="Allow mixed content (Enabling allows both secure and insecure content, images and scripts to render and execute. Disabling allows only secure content.)"
                    checked={ this.state.disablewebsecurity } onChange={ this.handleChangeDisableWebSecurity } />);
-    //OSX has an option in the tray, to set the app to autostart, so we choose to not support this option for OSX
+    //OSX has an option in the Dock, to set the app to autostart, so we choose to not support this option for OSX
     if (process.platform === 'win32' || process.platform === 'linux') {
       options.push(<Input key="inputAutoStart" ref="autostart" type="checkbox" label="Start app on login." checked={ this.state.autostart } onChange={ this.handleChangeAutoStart } />);
     }
@@ -179,55 +165,58 @@ var SettingsPage = React.createClass({
       </Row>
       ) : null;
 
-    var notificationSettings = [
-      {
-        label: 'Never',
-        state: 0
-      },
-      /* ToDo: Idle isn't implemented yet
-      {
-        label: 'Only when idle (after 10 seconds)',
-        state: 1
-      },*/
-      {
-        label: 'Always',
-        state: 2
-      }
-    ];
+    var notifications_row = null;
+    if (process.platform === 'win32') {
+      var notificationSettings = [
+        {
+          label: 'Never',
+          state: 0
+        },
+        /* ToDo: Idle isn't implemented yet
+        {
+          label: 'Only when idle (after 10 seconds)',
+          state: 1
+        },*/
+        {
+          label: 'Always',
+          state: 2
+        }
+      ];
 
-    var that = this;
-    var notificationElements = notificationSettings.map(function(item) {
-      var boundClick = that.handleFlashWindowSetting.bind(that, item);
-      return (
-        <Input key={ "flashWindow" + item.state } name="handleFlashWindow" ref={ "flashWindow" + item.state } type="radio" label={ item.label } value={ item.state } onChange={ boundClick }
-          checked={ that.state.notifications.flashWindow == item.state ? "checked" : "" } />
-        );
-    });
+      var that = this;
+      var notificationElements = notificationSettings.map(function(item) {
+        var boundClick = that.handleFlashWindowSetting.bind(that, item);
+        return (
+          <Input key={ "flashWindow" + item.state } name="handleFlashWindow" ref={ "flashWindow" + item.state } type="radio" label={ item.label } value={ item.state } onChange={ boundClick }
+            checked={ that.state.notifications.flashWindow == item.state ? "checked" : "" } />
+          );
+      });
 
-    var notifications = (
-    <Row>
-      <Col md={ 12 }>
-      <h2>Notifications</h2> Configure, that the taskicon in the taskbar blinks when you were mentioned.
-      { notificationElements }
-      </Col>
-    </Row>
-    )
+      notifications_row = (
+        <Row>
+          <Col md={ 12 }>
+          <h3>Notifications</h3> Configure, that the taskicon in the taskbar blinks when new message arrives.
+          { notificationElements }
+          </Col>
+        </Row>
+      );
+    }
 
     return (
       <Grid className="settingsPage">
         <Row>
-          <Col xs={ 4 } sm={ 1 } md={ 2 } lg={ 2 }>
+          <Col xs={ 4 } sm={ 2 } md={ 2 } lg={ 1 }>
           <h2>Teams</h2>
           </Col>
-          <Col xs={ 4 } sm={ 2 } md={ 1 } lg={ 1 } mdPull={ 1 }>
-          <Button className="pull-right" style={ buttonStyle } bsSize="small" onClick={ this.handleShowTeamForm }>
+          <Col xs={ 8 } sm={ 10 } md={ 10 } lg={ 11 }>
+          <Button bsSize="small" style={ { marginTop: 20 } } onClick={ this.toggleShowTeamForm }>
             <Glyphicon glyph="plus" />
           </Button>
           </Col>
         </Row>
         { teams_row }
         { options_row }
-        { notifications }
+        { notifications_row }
         <div>
           <hr />
         </div>
@@ -264,7 +253,7 @@ var TeamList = React.createClass({
     var teams = this.props.teams;
 
     // check if team already exists and then change existing team or add new one
-    if (!team.index && teams[team.index]) {
+    if ((team.index !== undefined) && teams[team.index]) {
       teams[team.index].name = team.name;
       teams[team.index].url = team.url;
     } else {