+++ test/modules/environment.js
@@ -1,7 +1,14 @@
 'use strict';
 
+const chai = require('chai');
+const chaiAsPromised = require('chai-as-promised');
+
+chai.should();
+chai.use(chaiAsPromised);
+
+
 const path = require('path');
-const webdriverio = require('webdriverio');
+const Application = require('spectron').Application;
 
 const source_root_dir = path.join(__dirname, '../..');
 const electron_binary_path = (function() {
@@ -13,29 +20,27 @@ const electron_binary_path = (function() {
     return path.join(source_root_dir, 'node_modules/electron-prebuilt/dist/electron' + exe_extension);
   }
 })();
-const config_file_path = path.join(source_root_dir, 'test_config.json');
+const config_file_path = path.join(source_root_dir, 'test/test_config.json');
 const mattermost_url = 'http://example.com/team';
-
-var options = {
-  host: 'localhost', // Use localhost as chrome driver server
-  port: 9515, // "9515" is the port opened by chrome driver.
-  desiredCapabilities: {
-    browserName: 'chrome',
-    chromeOptions: {
-      binary: electron_binary_path, // Path to your Electron binary.
-      args: ['app=' + path.join(source_root_dir, 'dist'), '--config-file=' + config_file_path] // Optional, perhaps 'app=' + /path/to/your/app/
-    }
-  }
-};
 
 module.exports = {
   sourceRootDir: source_root_dir,
   configFilePath: config_file_path,
   mattermostURL: mattermost_url,
-  spawnChromeDriver: function() {
-    return require('child_process').spawn('node_modules/chromedriver/lib/chromedriver/chromedriver', ['--url-base=wd/hub', '--port=9515']);
+  getSpectronApp: function() {
+    const app = new Application({
+      path: electron_binary_path,
+      args: [`${path.join(source_root_dir, 'dist')}`, '--config-file=' + config_file_path]
+    });
+    chaiAsPromised.transferPromiseness = app.transferPromiseness
+    return app;
   },
-  getWebDriverIoClient: function() {
-    return webdriverio.remote(options);
+  shouldTestForPlatforms: function(testCase, platforms) {
+    if (platforms.indexOf(process.platform) !== -1) {
+      return;
+    }
+    else {
+      testCase.skip();
+    }
   }
 }